Change var name. Also removed undefined checks
Removed undefined checks because the variable is always
going to be string. This is to avoid this:

"Warning: A component is changing an uncontrolled input
of type number to be controlled."

Having the state variable makes the variable an
uncrontolled input. By making it a string it is a
controlled input (controlled by the React engine
and not the DOM).